#27 [feat] : 채팅 대화 기록 조회하기 API 연동, diaryId 넘겨주기 구현

+++ app/src/main/java/com/konkuk/strhat/feature/diary/navigation/DiaryNavigation.kt
@@ -62,7 +62,7 @@ fun NavGraphBuilder.diaryNavGraph(
     onNavigateToMyPage: () -> Unit,
     onNavigateToTodayStressScore: (String) -> Unit,
     onPopBackStack: () -> Unit,
-    onNavigateToMyPageChatHistory: () -> Unit,
+    onNavigateToMyPageChatHistory: (Int) -> Unit,
     navController: NavController
 ) {
     composable<MainTabRoute.Diary> {
